Bumped minor version
Potential breaking changes:
The filename of the main file has been updated to something more
sensible for use in a web browser. The package.json file has been
updated, so this should not affect people using npm in the normal way.

New features:
The library can now be used directly in web browsers